langchain-with-krish/requirements.txt
langchain
python-dotenv
langchain_community
pypdf
bs4
langchain_text_splitters
langchain-openai
sentence-transformers
langchain-huggingface
chromadb
faiss-cpu
langchain-groq
streamlit